refactor register callbacks into multiple files for readability and good code structure
