+++ src/datadoc_editor/app.py
@@ -20,6 +20,7 @@ from flask_healthz import healthz
 
 from datadoc_editor import config
 from datadoc_editor import state
+from datadoc_editor.frontend.callbacks import register_all_callbacks
 from datadoc_editor.frontend.callbacks.register_callbacks import register_callbacks
 from datadoc_editor.frontend.components.control_bars import build_controls_bar
 from datadoc_editor.frontend.components.control_bars import build_footer_control_bar
@@ -88,7 +89,7 @@ def build_app(app: Dash) -> Dash:
         className="app-wrapper",
     )
 
-    register_callbacks(app)
+    register_all_callbacks(app)
 
     return app
