+++ src/datadoc_editor/frontend/callbacks/__init__.py
@@ -6,3 +6,10 @@ testing because the decorated functions are difficult to test.
 
 The functions where the logic lives should be categorised into files.
 """
+
+from .register_callbacks import register_callbacks
+from .pseudonymization_callbacks import register_pseudonymization_callbacks
+
+def register_all_callbacks(app):
+    register_callbacks(app)
+    register_pseudonymization_callbacks(app)
